requirements.txt
# Framework principal
streamlit==1.35.0

# Análisis de datos para reportes
pandas==2.1.4

# Base de datos
# SQLite viene incluido con Python, no necesita instalación

# Requests para webhooks a n8n y Mercado Pago
requests==2.31.0

# Utilidades
python-dotenv==1.0.0

Pillow==10.1.0

psycopg2-binary==2.9.9
python-dotenv==1.0.0